requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
paddleocr==2.7.0.3
paddlepaddle==2.5.1
numpy==1.24.3
python-dotenv==1.0.0